Extract conflict detection; unify the inverse-leg check
- Move replay-equality and reversal-link guards (and LedgerConflictError)
  out of store.ts into a cohesive conflicts.ts, bringing store.ts to 136
  lines.
- Add a pure isInverseOf to the ledger and use it for both building a
  reversal (reverseOf) and verifying one before insert (assertReverses),
  so the "exact inverse" rule lives in one place. Tested both ways,
  including that reverseOf's output satisfies isInverseOf.

100% coverage, 0% duplication, lint and typecheck clean.
