+++ src/shared/ledger/reverse.ts
@@ -1,6 +1,30 @@
 /** Pure construction of a reversing transfer (admin void/correction). */
 
-import type { Transfer, TransferInput } from "./types.ts";
+import { sameAccount } from "./account.ts";
+import type { AccountRef, Transfer, TransferInput } from "./types.ts";
+
+/** The fields that decide whether one leg exactly undoes another. */
+type DirectedAmount = {
+  readonly amount: number;
+  readonly currency: string;
+  readonly source: AccountRef;
+  readonly destination: AccountRef;
+};
+
+/**
+ * True when `leg` exactly undoes `original`: same amount and currency, with
+ * source and destination swapped. A freshly built reversal (see {@link reverseOf})
+ * satisfies this by construction; the store also checks a leg carrying a
+ * `reversesId` against it before inserting, so a bad link can't void nothing.
+ */
+export const isInverseOf = (
+  leg: DirectedAmount,
+  original: DirectedAmount,
+): boolean =>
+  leg.amount === original.amount &&
+  leg.currency === original.currency &&
+  sameAccount(leg.source, original.destination) &&
+  sameAccount(leg.destination, original.source);
 
 /**
  * Metadata the caller supplies for a reversal — the ledger reads no clock and
