+++ test/lib/ledger/reverse.test.ts
@@ -2,7 +2,7 @@ import { expect } from "@std/expect";
 import { describe, it } from "@std/testing/bdd";
 import { account } from "#shared/ledger/account.ts";
 import { balanceOf } from "#shared/ledger/project.ts";
-import { reverseOf } from "#shared/ledger/reverse.ts";
+import { isInverseOf, reverseOf } from "#shared/ledger/reverse.ts";
 import { makeTransfer } from "./factory.ts";
 
 const meta = {
@@ -51,5 +51,26 @@ describe("reverseOf", () => {
     const reversal = makeTransfer({ ...reverseOf(original, meta), id: 8 });
     expect(balanceOf(revenue)([original])).toBe(5000);
     expect(balanceOf(revenue)([original, reversal])).toBe(0);
+  });
+});
+
+describe("isInverseOf", () => {
+  const original = makeTransfer({
+    amount: 5000,
+    currency: "GBP",
+    destination: account("revenue", 45),
+    source: account("attendee", 88),
+  });
+
+  it("accepts the leg that reverseOf builds", () => {
+    expect(isInverseOf(reverseOf(original, meta), original)).toBe(true);
+  });
+
+  it("rejects a different amount, currency, or direction", () => {
+    const inverse = reverseOf(original, meta);
+    expect(isInverseOf({ ...inverse, amount: 4000 }, original)).toBe(false);
+    expect(isInverseOf({ ...inverse, currency: "USD" }, original)).toBe(false);
+    // Same direction as the original (ends not swapped).
+    expect(isInverseOf(original, original)).toBe(false);
   });
 });
